requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==5.2.1
django-ninja==1.4.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
idna==3.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
